Fixed notification issue
We were not getting the new id of inserted application which resulted
into creating faulty notification work, this was the reason why on
clicking the notification that same application was not getting opened,
and also the cancellation of works was not happening.
Minor issue of scrolling applications list also fixed. New behavior,
toolbar only scrolls if applications list is long enough.